requirements.txt
streamlit
ollama
pillow